rename timezones.ts to timezone-info.ts
